anupk-adda-r42-1c6b12bd4054dc55b3d549a410c0c6ec40a7f6d3/agent-service/requirements.txt
# LangChain and LangGraph
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
httpx>=0.26.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
